requirements.txt
django==4.2
environs==9.5.0
django-phonenumber-field==7.0.2
phonenumbers==8.13.10
pillow==9.5.0
python-dateutil==2.8.2